requirements.txt
defusedxml==0.7.1
fonttools==4.60.1
fpdf2==2.8.4
pillow==11.3.0
